More barrel-busting to fix broken test
