+++ packages/synapse-react-client/src/components/GenericCard/TableRowGenericCard.tsx
@@ -35,7 +35,7 @@ import CitationPopover from '../CitationPopover'
 import { EntityDownloadConfirmation } from '../EntityDownloadConfirmation'
 import { HeaderCardVariant } from '../HeaderCard'
 import IconList from '../IconList'
-import { useQueryContext } from '../QueryContext'
+import { useQueryContext } from '../QueryContext/QueryContext'
 import { useQueryVisualizationContext } from '../QueryVisualizationWrapper'
 import GenericCard from './GenericCard'
 import GenericCardActionButton from './GenericCardActionButton'
