+++ packages/synapse-react-client/src/components/GenericCard/TableRowGenericCard.vitest.tsx
@@ -1,5 +1,3 @@
-import { CardLink } from '@/components/CardContainer/CardLink'
-import { getLinkParams } from '@/components/GenericCard/CardUtils'
 import PortalDOI from '@/components/GenericCard/PortalDOI/PortalDOI'
 import {
   getCandidateDoiId,
@@ -26,10 +24,9 @@ import { act, fireEvent, render, screen, within } from '@testing-library/react'
 import { cloneDeep } from 'lodash-es'
 import { mockAllIsIntersecting } from 'react-intersection-observer/test-utils'
 import { beforeEach, describe, it, test, vi } from 'vitest'
-import { TargetEnum } from '@/utils/html/TargetEnum'
 import { EntityDownloadConfirmation } from '../EntityDownloadConfirmation'
 import { QueryVisualizationWrapper } from '../QueryVisualizationWrapper'
-import QueryWrapper from '../QueryWrapper'
+import { QueryWrapper } from '../QueryWrapper/QueryWrapper'
 import { FileHandleLink } from '../widgets/FileHandleLink'
 import { ImageFileHandle } from '../widgets/ImageFileHandle'
 import { CARD_SHORT_DESCRIPTION_CSS } from './CollapsibleDescription'
@@ -384,150 +381,6 @@ describe('TableRowGenericCard tests', () => {
         },
         expect.anything(),
       )
-    })
-  })
-
-  describe('it makes the correct URL for the title', () => {
-    test('creates a link to synapse', () => {
-      const synId = 'syn12345678'
-      const synLink = `https://www.synapse.org/Synapse:${synId}`
-      const { href, target } = getLinkParams(
-        synId,
-        undefined,
-        undefined,
-        undefined,
-      )
-      expect(href).toEqual(synLink)
-      expect(target).toEqual(TargetEnum.CURRENT_WINDOW)
-    })
-
-    test('creates a DOI link', () => {
-      const doi = '10.1093/neuonc/noy046'
-      const doiLink = `https://dx.doi.org/${doi}`
-      const { href, target } = getLinkParams(
-        doi,
-        undefined,
-        undefined,
-        undefined,
-      )
-      expect(href).toEqual(doiLink)
-      expect(target).toEqual(TargetEnum.NEW_WINDOW)
-    })
-
-    test('creates a DOI link PORTALS-1801', () => {
-      const doi = '10.1007/s00401-020-02230-x '
-      const doiLink = `https://dx.doi.org/${doi.trim()}`
-      const { href, target } = getLinkParams(
-        doi,
-        undefined,
-        undefined,
-        undefined,
-      )
-      expect(href).toEqual(doiLink)
-      expect(target).toEqual(TargetEnum.NEW_WINDOW)
-    })
-
-    test('creates an internal parameterized link', () => {
-      const value = '1234'
-      const data = [value]
-      const URLColumnName = 'Grant Number'
-      const matchColumnName = 'Funder'
-      const schema = {
-        [matchColumnName]: 0,
-      }
-      const titleLinkConfig: CardLink = {
-        isMarkdown: false,
-        baseURL: 'Explore/Projects',
-        matchColumnName,
-        URLColumnName,
-      }
-      const expectedLink = `/${titleLinkConfig.baseURL}?${URLColumnName}=${value}`
-      const { href, target } = getLinkParams('', titleLinkConfig, data, schema)
-      expect(href).toEqual(expectedLink)
-      expect(target).toEqual(TargetEnum.CURRENT_WINDOW)
-    })
-
-    test('creates an internal details page link', () => {
-      const value = '1234+5+6'
-      const data = [value]
-      const URLColumnName = 'Grant Number'
-      const matchColumnName = 'Funder'
-      const schema = {
-        [matchColumnName]: 0,
-      }
-      const titleLinkConfig: CardLink = {
-        isMarkdown: false,
-        baseURL: 'Explore/Programs/DetailsPage',
-        matchColumnName,
-        URLColumnName,
-      }
-      const expectedLink = `/${
-        titleLinkConfig.baseURL
-      }?${URLColumnName}=${encodeURIComponent(value)}`
-      const { href: href1, target: target1 } = getLinkParams(
-        '',
-        titleLinkConfig,
-        data,
-        schema,
-      )
-      expect(href1).toEqual(expectedLink)
-      // PORTALS-2254: Open DetailsPage links in a new window by default
-      expect(target1).toEqual(TargetEnum.NEW_WINDOW)
-
-      titleLinkConfig.target = TargetEnum.FULL_WINDOW_BODY
-      const { href: href2, target: target2 } = getLinkParams(
-        '',
-        titleLinkConfig,
-        data,
-        schema,
-      )
-      expect(href2).toEqual(expectedLink)
-      // PORTALS-2254: Verify we can override the target via the parameter in the CardLink config
-      expect(target2).toEqual(TargetEnum.FULL_WINDOW_BODY)
-    })
-
-    it('uses another column value for link override', () => {
-      const value = 'foo'
-      const hrefOverrideValue = 'bar'
-      const data = [value, hrefOverrideValue]
-      const matchColumnName = 'Funder'
-      const hrefOverrideColumnName = 'Override Col'
-      const schema = {
-        [matchColumnName]: 0,
-        [hrefOverrideColumnName]: 1,
-      }
-      const titleLinkConfig: CardLink = {
-        isMarkdown: false,
-        matchColumnName,
-        overrideLinkURLColumnName: hrefOverrideColumnName,
-      }
-      const expectedLink = hrefOverrideValue
-      const { href, target } = getLinkParams('', titleLinkConfig, data, schema)
-      expect(href).toEqual(expectedLink)
-      expect(target).toEqual(TargetEnum.CURRENT_WINDOW)
-    })
-
-    it('uses another column value for link override with transform', () => {
-      const value = 'foo'
-      const hrefOverrideValue = 'bar'
-      const data = [value, hrefOverrideValue]
-      const matchColumnName = 'Funder'
-      const hrefOverrideColumnName = 'Override Col'
-      const transform = (value: string) => `https://example.com/${value}`
-      const schema = {
-        [matchColumnName]: 0,
-        [hrefOverrideColumnName]: 1,
-      }
-      const titleLinkConfig: CardLink = {
-        isMarkdown: false,
-        matchColumnName,
-        overrideLinkURLColumnName: hrefOverrideColumnName,
-        overrideLinkURLColumnTransform: transform,
-      }
-      const expectedLink = 'https://example.com/bar'
-      const { href, target } = getLinkParams('', titleLinkConfig, data, schema)
-      expect(href).toEqual(expectedLink)
-      expect(target).toEqual(TargetEnum.CURRENT_WINDOW)
     })
   })
