+++ packages/synapse-react-client/src/components/SynapseTable/SynapseTableCell/SynapseTableCell.stories.tsx
@@ -1,7 +1,7 @@
 import { mockTableEntity } from '@/mocks/entity/mockTableEntity'
 import { ColumnTypeEnum } from '@sage-bionetworks/synapse-types'
 import { Meta, StoryObj } from '@storybook/react'
-import { QueryContextProvider } from '../../../index'
+import { QueryContextProvider } from '@/components/QueryContext/QueryContext'
 import SynapseTableCell, { SynapseTableCellProps } from './SynapseTableCell'
 
 const meta: Meta<SynapseTableCellProps> = {
