+++ packages/synapse-react-client/src/utils/hooks/useOneSageURL.ts
@@ -1,5 +1,5 @@
 import { useMemo } from 'react'
-import { useSynapseContext } from '../context'
+import { useSynapseContext } from '../context/SynapseContext'
 import { BackendDestinationEnum, getEndpoint } from '../functions/index'
 import {
   ONE_SAGE_APPID_QUERY_PARAM_KEY,
